cms-admin: Extract createActionLogDialogQueries factory
